+++ iModelCore/ECDb/ECDb/IssueReporter.cpp
@@ -743,6 +743,7 @@ IssueId ECDbIssueId::ECDb_0743 = IssueId("ECDb_0743");
 IssueId ECDbIssueId::ECDb_0744 = IssueId("ECDb_0744");
 IssueId ECDbIssueId::ECDb_0745 = IssueId("ECDb_0745");
 IssueId ECDbIssueId::ECDb_0746 = IssueId("ECDb_0746");
+IssueId ECDbIssueId::ECDb_0747 = IssueId("ECDb_0747");
 
 //---------------------------------------------------------------------------------------
 // @bsimethod