+++ iModelCore/ECDb/ECDb/ProfileUpgrader.h
@@ -70,7 +70,8 @@ BEGIN_BENTLEY_SQLITE_EC_NAMESPACE
 #define TABLEDDL_Feature "CREATE TABLE " TABLE_Feature "(" \
             "Name TEXT PRIMARY KEY NOT NULL COLLATE NOCASE," \
             "Description TEXT NOT NULL," \
-            "Compat TEXT NOT NULL CHECK(Compat IN ('Warn','ReadOnly','NoSchemaImport','Refuse')));"
+            "Compat TEXT NOT NULL," \
+            "Fallback TEXT NOT NULL COLLATE NOCASE CHECK(Fallback IN ('Warn','ReadOnly','NoSchemaImport','NoChangesetGeneration','Refuse')));"
 
 //=======================================================================================
 // @bsiclass